requirements.txt
crewai>=0.28.8
crewai[tools]>=0.1.2
google-generativeai>=0.5.4
python-dotenv>=1.0.1
chromadb>=0.4.24
moviepy==1.0.3
schedule>=1.2.1
google-api-python-client>=2.126.0
google-auth-httplib2>=0.2.0
google-auth-oauthlib>=1.2.0
langchain-google-genai
google-genai
langchain-google-vertexai
replicate
elevenlabs